Combine LoggingResolvers extensions into one file
