--- log4j-core-test/src/main/java/org/apache/logging/log4j/core/test/junit/LoggerResolver.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to you under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.logging.log4j.core.test.junit;
-
-import java.lang.reflect.Parameter;
-import org.apache.logging.log4j.core.Logger;
-import org.apache.logging.log4j.plugins.di.Keys;
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.jupiter.api.extension.ParameterContext;
-import org.junit.jupiter.api.extension.ParameterResolutionException;
-import org.junit.jupiter.api.extension.ParameterResolver;
-
-class LoggerResolver implements ParameterResolver {
-    @Override
-    public boolean supportsParameter(final ParameterContext parameterContext, final ExtensionContext extensionContext)
-            throws ParameterResolutionException {
-        return parameterContext.getParameter().getType().isAssignableFrom(Logger.class);
-    }
-
-    @Override
-    public Logger resolveParameter(final ParameterContext parameterContext, final ExtensionContext extensionContext)
-            throws ParameterResolutionException {
-        final Parameter parameter = parameterContext.getParameter();
-        final String loggerName = Keys.hasName(parameter)
-                ? Keys.getName(parameter)
-                : extensionContext.getRequiredTestClass().getCanonicalName();
-        return Log4jExtension.getRequiredLoggerContext(extensionContext).getLogger(loggerName);
-    }
-}
